Remove IcebergTaskWriters
